+++ runtime/calamari/Cargo.toml
@@ -92,6 +92,7 @@ manta-support = { package = "pallet-manta-support", path = "../../pallets/manta-
 pallet-asset-manager = { path = '../../pallets/asset-manager', default-features = false }
 pallet-manta-pay = { path = '../../pallets/manta-pay', default-features = false, features = ["runtime"] }
 pallet-manta-sbt = { path = '../../pallets/manta-sbt', default-features = false, features = ["runtime"] }
+pallet-name-service = { path = '../../pallets/name-service', default-features = false}
 pallet-parachain-staking = { path = '../../pallets/parachain-staking', default-features = false }
 pallet-tx-pause = { path = '../../pallets/tx-pause', default-features = false }
 runtime-common = { path = '../common', default-features = false }
