+++ runtime/dolphin/Cargo.toml
@@ -95,7 +95,6 @@ manta-support = { package = "pallet-manta-support", path = "../../pallets/manta-
 pallet-asset-manager = { path = '../../pallets/asset-manager', default-features = false }
 pallet-manta-pay = { path = '../../pallets/manta-pay', default-features = false, features = ["runtime"] }
 pallet-manta-sbt = { path = '../../pallets/manta-sbt', default-features = false, features = ["runtime"] }
-pallet-name-service ={ path = '../../pallets/name-service', default-features = false}
 pallet-tx-pause = { path = '../../pallets/tx-pause', default-features = false }
 runtime-common = { path = '../common', default-features = false }
 session-key-primitives = { path = '../../primitives/session-keys', default-features = false }
