+++ runtime/dolphin/src/lib.rs
@@ -56,7 +56,7 @@ use frame_system::{
 };
 use manta_primitives::{
     constants::{
-        time::*, RocksDbWeight, NAME_SERVICE_PALLET_ID, STAKING_PALLET_ID, TREASURY_PALLET_ID,
+        time::*, RocksDbWeight, STAKING_PALLET_ID, TREASURY_PALLET_ID,
         WEIGHT_PER_SECOND,
     },
     types::{AccountId, Balance, BlockNumber, Hash, Header, Index, Signature},
@@ -286,7 +286,6 @@ impl Contains<RuntimeCall> for BaseFilter {
             | RuntimeCall::MantaPay(_)
             | RuntimeCall::Preimage(_)
             | RuntimeCall::MantaSbt(_)
-            | RuntimeCall::NameService(_)
             | RuntimeCall::TransactionPause(_)
             | RuntimeCall::AssetManager(pallet_asset_manager::Call::update_outgoing_filtered_assets {..})
             | RuntimeCall::Utility(_) => true,
@@ -401,17 +400,6 @@ impl pallet_utility::Config for Runtime {
 impl pallet_sudo::Config for Runtime {
     type RuntimeEvent = RuntimeEvent;
     type RuntimeCall = RuntimeCall;
-}
-
-parameter_types! {
-    pub const NameServicePalletId: PalletId = NAME_SERVICE_PALLET_ID;
-}
-
-impl pallet_name_service::Config for Runtime {
-    type RuntimeEvent = RuntimeEvent;
-    type PalletId = NameServicePalletId;
-    type RegisterWaitingPeriod = ConstU32<2>;
-    //type WeightInfo = weights::pallet_name_service::SubstrateWeight<Runtime>;
 }
 
 parameter_types! {
@@ -780,7 +768,6 @@ construct_runtime!(
         AssetManager: pallet_asset_manager::{Pallet, Call, Storage, Config<T>, Event<T>} = 46,
         MantaPay: pallet_manta_pay::{Pallet, Call, Storage, Event<T>} = 47,
         MantaSbt: pallet_manta_sbt::{Pallet, Call, Storage, Event<T>} = 48,
-        NameService: pallet_name_service::{Pallet, Call, Storage, Event<T>} = 49,
     }
 );
