+++ runtime/dolphin/src/weights/mod.rs
@@ -27,7 +27,6 @@ pub mod pallet_collective;
 pub mod pallet_democracy;
 pub mod pallet_manta_pay;
 pub mod pallet_manta_sbt;
-//pub mod pallet_name_service;
 pub mod pallet_membership;
 pub mod pallet_multisig;
 pub mod pallet_preimage;
