+++ runtime/dolphin/src/weights/pallet_manta_sbt.rs
@@ -58,7 +58,7 @@ pub trait WeightInfo {
 
 /// Weights for pallet_manta_sbt using the Substrate node and recommended hardware.
 pub struct SubstrateWeight<T>(PhantomData<T>);
-impl<T: frame_system::Config>WeightInfo for SubstrateWeight<T> {
+impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
 	// Storage: MantaSbt ReservedIds (r:1 w:1)
 	// Storage: MantaSbt UtxoSet (r:1 w:1)
 	// Storage: MantaSbt ShardTrees (r:1 w:1)
